requirements.txt
flask==2.2.5 
flask_sqlalchemy==3.1.1 
flask-migrate==4.0.7
psycopg2-binary==2.9.3 
PyJWT==2.3.0
python-dotenv==1.0.1
flask-swagger-ui==4.11.1
pytest==8.3.2 
pytest-flask==1.3.0